This has the code for gravity, magnetism
The problem with magnetism previously was: (1) I was trying to be cute with advection by doing things in reverse, (2) The singularity of distance needed mitigation.
These changes allowed me to implement gravity as well (pointing to the cursor). This also provides for allowing a compressible fluid where the most dense part of the fluid interact most strongly with gravity and is allowed to move inward (which would otherwise violate "gradient p=0" of the Navier-stokes equations for an incompressible fluid).